+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for json-query
 
+## 0.1.1.0 -- ???
+
+* Add `Monad` implementation for `MemberParser`.
+* Add `Alternative` implementation for `Parser` and `MemberParser`.
+
 ## 0.1.0.0 -- 2021-03-22
 
 * Initial release.